feat: shared data keys implementation

+++ flutter/iot_receiver/lib/widgets/hro2_drawer_widget.dart
@@ -1,5 +1,9 @@
+import 'package:at_client_mobile/at_client_mobile.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:iot_receiver/screens/data_owners_screen.dart';
 import 'package:iot_receiver/screens/onboarding_screen.dart';
+import 'package:iot_receiver/widgets/new_data_owner_dialog.dart';
 import '../services/hro2_data_service.dart';
 import '../screens/devices_screen.dart';
 import '../screens/home_screen.dart';
@@ -7,9 +11,17 @@ import '../screens/receivers_screen.dart';
 import 'new_device_dialog.dart';
 import 'new_receiver_dialog.dart';
 
-class HRo2DrawerWidget extends StatelessWidget {
+class HRo2DrawerWidget extends StatefulWidget {
   const HRo2DrawerWidget({Key? key}) : super(key: key);
 
+  @override
+  State<HRo2DrawerWidget> createState() => _HRo2DrawerWidgetState();
+}
+
+class _HRo2DrawerWidgetState extends State<HRo2DrawerWidget> {
+  final bool isAdmin =
+      AtClientManager.getInstance().atClient.getCurrentAtSign() ==
+          "@mwcmanager";
   @override
   Widget build(BuildContext context) {
     return Drawer(
@@ -21,7 +33,7 @@ class HRo2DrawerWidget extends StatelessWidget {
             decoration: BoxDecoration(
               color: Colors.white,
             ),
-            child: Text('HRo2 Options'),
+            child: Text('HRO2 Options'),
           ),
           ListTile(
             title: const Text('See HRo2 readings'),
@@ -29,40 +41,66 @@ class HRo2DrawerWidget extends StatelessWidget {
               Navigator.of(context).pushNamed(HomeScreen.id);
             },
           ),
-          ListTile(
-            title: const Text('See all devices'),
-            onTap: () {
-              Navigator.of(context).pushNamed(DevicesScreen.id);
-            },
-          ),
-          ListTile(
-            title: const Text('Add a new device'),
-            onTap: () {
-              Navigator.of(context).pushNamed(NewHrO2Device.id);
-            },
-          ),
-          ListTile(
-            title: const Text('See all receivers'),
-            onTap: () {
-              Navigator.of(context).pushNamed(ReceiversScreen.id);
-            },
-          ),
-          ListTile(
-            title: const Text('Add a new receiver'),
-            onTap: () {
-              Navigator.of(context).pushNamed(NewHrO2Receiver.id);
-            },
-          ),
+          if (isAdmin)
+            ListTile(
+              title: const Text('See all devices'),
+              onTap: () {
+                Navigator.of(context).pushNamed(DevicesScreen.id);
+              },
+            ),
+          if (isAdmin)
+            ListTile(
+              title: const Text('Add a new device'),
+              onTap: () {
+                Navigator.of(context).pushNamed(NewHrO2Device.id);
+              },
+            ),
+          if (!isAdmin)
+            ListTile(
+              title: const Text('See all receivers'),
+              onTap: () {
+                Navigator.of(context).pushNamed(ReceiversScreen.id);
+              },
+            ),
+          if (!isAdmin)
+            ListTile(
+              title: const Text('Add a new receiver'),
+              onTap: () {
+                Navigator.of(context).pushNamed(NewHrO2Receiver.id);
+              },
+            ),
+          if (isAdmin)
+            ListTile(
+              title: const Text('Add a new data owner'),
+              onTap: () {
+                Navigator.of(context).pushNamed(NewHrO2DataOwner.id);
+              },
+            ),
+          if (isAdmin)
+            ListTile(
+              title: const Text('See all data owners'),
+              onTap: () {
+                Navigator.of(context).pushNamed(DataOwnersScreen.id);
+              },
+            ),
           ListTile(
             title: const Text('Change atSign'),
             onTap: () async {
               Navigator.of(context).pushNamed(OnboardingScreen.id);
             },
           ),
+          if (isAdmin)
+            ListTile(
+              title: const Text('Reset app data'),
+              onTap: () async {
+                Hro2DataService().deleteAllData();
+                Navigator.of(context).pop();
+              },
+            ),
           ListTile(
-            title: const Text('Reset app data'),
-            onTap: () async {
-              await Hro2DataService().deleteAllData();
+            title: const Text('Close'),
+            onTap: () {
+              SystemNavigator.pop();
             },
           ),
         ],
